Working CRUD app with rooms, small fixes in movies

+++ ticket-service/src/main/java/com/epam/training/ticketservice/commandprompt/commands/MovieCommand.java
@@ -21,18 +21,21 @@ public class MovieCommand {
     }
 
     @ShellMethod(key = "update movie",value = "Update a movie")
-    public Movie updateMovie(String title,String genre, Integer lengthInMinutes) {
+    public Movie updateMovie(String title,String genre, Integer lengthInMinutes) throws Exception {
+        if (movieService.getMovieByTitle(title).isEmpty()) throw new Exception("There is no movie with that title");
         movieService.updateMovie(title, genre, lengthInMinutes);
         return new Movie(title, genre, lengthInMinutes);
     }
 
     @ShellMethod(key = "delete movie",value = "Delete a movie")
-    public void deleteMovie(String title) {
+    public void deleteMovie(String title) throws Exception {
+        if (movieService.getMovieByTitle(title).isEmpty()) throw new Exception("There is no movie with that title");
         movieService.deleteMovie(title);
     }
 
     @ShellMethod(key = "list movies",value = "List of movies")
-    public List<Movie> listMovies() {
+    public List<Movie> listMovies() throws Exception {
+        if (movieService.getMovieList()==null) throw new Exception("There are no movies at the moment");
         return movieService.getMovieList();
     }
 }
